update: refactor all of api course

+++ src/main/java/vn/edu/ut/service/impl/CoursesServiceImpl.java
@@ -2,27 +2,25 @@ package vn.edu.ut.service.impl;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import vn.edu.ut.entity.Category;
 import vn.edu.ut.entity.CourseInfo;
 import vn.edu.ut.entity.Courses;
 import vn.edu.ut.entity.Lesson;
-import vn.edu.ut.entity.Review;
 import vn.edu.ut.entity.Video;
+import vn.edu.ut.enums.ErrorCode;
 import vn.edu.ut.enums.InformationType;
 import vn.edu.ut.enums.LessonType;
-import vn.edu.ut.exception.AppException;
+import vn.edu.ut.exception.AppApiException;
 import vn.edu.ut.exception.ResourceNotFoundException;
 import vn.edu.ut.payload.ClassResponse;
-import vn.edu.ut.payload.chapter.ChapterDto;
 import vn.edu.ut.payload.chapter.ChapterReturnDetailResponse;
 import vn.edu.ut.payload.course.CourseInfoRequest;
 import vn.edu.ut.payload.course.CourseResponse;
@@ -30,7 +28,6 @@ import vn.edu.ut.payload.course.CourseReturnDetailPageResponse;
 import vn.edu.ut.payload.course.CourseReturnHomePageResponse;
 import vn.edu.ut.payload.course.CourseReturnSearch;
 import vn.edu.ut.payload.course.CoursesRequest;
-import vn.edu.ut.payload.lesson.LessonResponse;
 import vn.edu.ut.payload.lesson.LessonReturnDetailResponse;
 import vn.edu.ut.repository.CategoryRepository;
 import vn.edu.ut.repository.CoursesRepository;
@@ -38,15 +35,12 @@ import vn.edu.ut.repository.LessonRepository;
 import vn.edu.ut.repository.VideoRepository;
 import vn.edu.ut.service.ICoursesService;
 import vn.edu.ut.utils.UploadFile;
-import vn.edu.ut.utils.Utils;
 
 import java.time.Duration;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -62,35 +56,31 @@ public class CoursesServiceImpl implements ICoursesService {
 
     @Override
     public CourseResponse createCourse(CoursesRequest coursesRequest, MultipartFile image) {
-        if (coursesRepository.existsCoursesByTitle(coursesRequest.getTitle())) {
-            throw new AppException(HttpStatus.BAD_REQUEST, "Tên khóa học đã được tạo trước đó!");
-        }
-
-        if (coursesRepository.existsCoursesBySlug(coursesRequest.getSlug())) {
-            throw new AppException(HttpStatus.BAD_REQUEST, "Slug khóa học đã được tạo trước đó!");
-        }
-
         Category category = categoryRepository.findById(coursesRequest.getCategoryId())
                 .orElseThrow(() -> new ResourceNotFoundException("Category", "id", coursesRequest.getCategoryId()));
 
-        Courses courses = new Courses();
+        if (coursesRepository.existsCoursesByTitle(coursesRequest.getTitle())) {
+            throw new AppApiException(ErrorCode.COURSE_NAME_EXISTED);
+        }
 
-        convertSomeAttributeToEntity(courses, coursesRequest);
+        if (coursesRepository.existsCoursesBySlug(coursesRequest.getSlug())) {
+            throw new AppApiException(ErrorCode.COURSE_SLUG_EXISTED);
+        }
 
+        Courses courses = new Courses(coursesRequest);
         String thumbnail = uploadFile.uploadFileOnCloudinary(image);
         courses.setThumbnail(thumbnail);
-
         courses.setCategory(category);
 
         for (CourseInfoRequest request : coursesRequest.getInfoList()) {
             courses.addInfoList(request.getValue(), InformationType.valueOf(request.getType()));
         }
+        coursesRepository.save(courses);
 
-        Courses savedCourse = coursesRepository.save(courses);
-
-        return modelMapper.map(savedCourse, CourseResponse.class);
+        return modelMapper.map(courses, CourseResponse.class);
     }
 
+    /*TODO: calculate total review and average review*/
     @Override
     public ClassResponse getAll(int pageNo, int pageSize, String sortBy, String sortDir, String keyword, Integer categoryId) {
         Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name())
@@ -98,39 +88,27 @@ public class CoursesServiceImpl implements ICoursesService {
 
         Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
 
-        Page<Courses> courses = null;
-        if (keyword != null && !keyword.isEmpty()) {
-            if (categoryId != null && categoryId > 0) {
-                courses = coursesRepository.searchInCategory(keyword, categoryId, pageable);
+        Page<Courses> coursesResult;
+        if (StringUtils.isNotEmpty(keyword)) {
+            if (categoryId != null) {
+                coursesResult = coursesRepository.searchInCategory(keyword, categoryId, pageable);
             } else {
-                courses = coursesRepository.search(keyword, pageable);
+                coursesResult = coursesRepository.search(keyword, pageable);
             }
         } else {
-            courses = coursesRepository.findAllInCategory(categoryId, pageable);
+            if (categoryId != null) {
+                coursesResult = coursesRepository.findAllInCategory(categoryId, pageable);
+            } else {
+                coursesResult = coursesRepository.findAll(pageable);
+            }
         }
+        List<Courses> listCourses = coursesResult.getContent();
 
-        if (keyword == null && categoryId == null) {
-            courses = coursesRepository.findAll(pageable);
-        }
-        List<Courses> listCourses = courses.getContent();
-
-        List<CourseResponse> content = listCourses.stream()
-                .map(course -> {
-                            CourseResponse courseResponse = modelMapper.map(course, CourseResponse.class);
-                            int totalReview = course.getListReviews().size();
-                            int totalRating = course.getListReviews().stream().mapToInt(Review::getRating).sum();
-                            double averageRating = (double) totalRating / totalReview;
-                            averageRating = Math.round(averageRating * 10.0) / 10.0;
-                            courseResponse.setTotalReview(totalReview);
-                            courseResponse.setAverageReview(averageRating);
-                            courseResponse.setChapterList(null);
-                            courseResponse.setInfoList(null);
-                            return courseResponse;
-                        }
-                )
+        List<CourseReturnHomePageResponse> content = listCourses.stream()
+                .map(courses -> modelMapper.map(courses, CourseReturnHomePageResponse.class))
                 .collect(Collectors.toList());
 
-        return ClassResponse.convertToClassResponse(courses, content);
+        return ClassResponse.convertToClassResponse(coursesResult, content);
     }
 
     @Override
@@ -139,7 +117,7 @@ public class CoursesServiceImpl implements ICoursesService {
                 .orElseThrow(() -> new ResourceNotFoundException("Course", "id", courseId));
 
         CourseResponse response = modelMapper.map(course, CourseResponse.class);
-        sortChapterAndLesson(response);
+//        sortChapterAndLesson(response);
         return response;
     }
 
@@ -151,13 +129,19 @@ public class CoursesServiceImpl implements ICoursesService {
         Category categoryInDB = categoryRepository.findById(coursesRequest.getCategoryId())
                 .orElseThrow(() -> new ResourceNotFoundException("Category", "id", coursesRequest.getCategoryId()));
 
-        Courses courses = coursesRepository.findByTitleOrSlug(coursesRequest.getTitle(), coursesRequest.getSlug());
-
-        if (courses != null) {
-            if (!Objects.equals(courses.getId(), courseInDB.getId())) {
-                throw new AppException(HttpStatus.BAD_REQUEST, "Tên/Slug khóa học đã tồn tại trước đó");
-            }
+        if (!coursesRepository.existsCoursesByTitleOrSlugAndId(coursesRequest.getTitle(),
+                coursesRequest.getSlug(), courseId)) {
+            throw new AppApiException(ErrorCode.COURSE_SLUG_NAME_EXISTED);
         }
+        courseInDB.setTitle(coursesRequest.getTitle());
+        courseInDB.setSlug(coursesRequest.getSlug());
+        courseInDB.setDescription(coursesRequest.getDescription());
+        courseInDB.setCategory(categoryInDB);
+        courseInDB.setPrice(coursesRequest.getPrice());
+        courseInDB.setDiscount(coursesRequest.getDiscount());
+        courseInDB.setEnabled(coursesRequest.isEnabled());
+        courseInDB.setPublished(coursesRequest.isPublished());
+        courseInDB.setFinished(coursesRequest.isFinished());
 
         if (img != null) {
             uploadFile.deleteImageInCloudinary(courseInDB.getThumbnail());
@@ -165,14 +149,9 @@ public class CoursesServiceImpl implements ICoursesService {
             courseInDB.setThumbnail(url);
         }
 
-        convertSomeAttributeToEntity(courseInDB, coursesRequest);
-
-        courseInDB.setCategory(categoryInDB);
-
         List<CourseInfo> infoList = new ArrayList<>();
-
         for (CourseInfoRequest request : coursesRequest.getInfoList()) {
-            CourseInfo info = null;
+            CourseInfo info;
             if (request.getId() != null) {
                 info = new CourseInfo(request.getId(), request.getValue(), InformationType.valueOf(request.getType()), courseInDB);
             } else {
@@ -180,50 +159,35 @@ public class CoursesServiceImpl implements ICoursesService {
             }
             infoList.add(info);
         }
-
         courseInDB.setInfoList(infoList);
+        coursesRepository.save(courseInDB);
 
-        Courses savedCourse = coursesRepository.save(courseInDB);
-
-        return modelMapper.map(savedCourse, CourseResponse.class);
+        return modelMapper.map(courseInDB, CourseResponse.class);
     }
 
     @Override
-    public String delete(Integer courseId) {
+    public void delete(Integer courseId) {
         Courses courseInDB = coursesRepository.findById(courseId)
                 .orElseThrow(() -> new ResourceNotFoundException("Course", "id", courseId));
 
         uploadFile.deleteImageInCloudinary(courseInDB.getThumbnail());
 
         coursesRepository.delete(courseInDB);
-
-        return "Xóa khóa học thành công";
     }
 
+    /*TODO: calculate total review and average review*/
     @Override
     public List<CourseReturnHomePageResponse> getCourseIntoHomePage(Integer categoryId) {
-        List<Courses> listCourses = null;
+        List<Courses> listCourses;
 
         if (categoryId == null) {
             listCourses = coursesRepository.findAll();
         } else {
-            Category category = categoryRepository.findById(categoryId)
-                    .orElseThrow(() -> new ResourceNotFoundException("Category", "id", categoryId));
-
             listCourses = coursesRepository.findAllByCategoryId(categoryId);
         }
-
         return listCourses.stream()
-                .map(courses -> {
-                    CourseReturnHomePageResponse response = modelMapper.map(courses, CourseReturnHomePageResponse.class);
-                    int totalReview = courses.getListReviews().size();
-                    int totalRating = courses.getListReviews().stream().mapToInt(Review::getRating).sum();
-                    double averageRating = (double) totalRating / totalReview;
-                    averageRating = Math.round(averageRating * 10.0) / 10.0;
-                    response.setTotalReview(totalReview);
-                    response.setAverageReview(averageRating);
-                    return response;
-                }).toList();
+                .map(courses -> modelMapper.map(courses, CourseReturnHomePageResponse.class))
+                .toList();
     }
 
     @Override
@@ -237,78 +201,52 @@ public class CoursesServiceImpl implements ICoursesService {
     }
 
     @Override
-    public String updateIsEnabled(Integer courseId, boolean isEnabled) {
-        Courses courses = coursesRepository.findById(courseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Courses", "id", courseId));
-
-        coursesRepository.switchEnabled(courseId, isEnabled);
-
-        return "SUCCESS";
+    public void updateIsEnabled(Integer courseId, boolean isEnabled) {
+        if (coursesRepository.existsById(courseId))
+            coursesRepository.switchEnabled(courseId, isEnabled);
+        else
+            throw new ResourceNotFoundException("Course", "id", courseId);
     }
 
     @Override
-    public String updateIsPublished(Integer courseId, boolean isPublished) {
-        Courses courses = coursesRepository.findById(courseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Courses", "id", courseId));
-
-        coursesRepository.switchPublished(courseId, isPublished);
-        return "SUCCESS";
+    public void updateIsPublished(Integer courseId, boolean isPublished) {
+        if (coursesRepository.existsById(courseId))
+            coursesRepository.switchPublished(courseId, isPublished);
+        else
+            throw new ResourceNotFoundException("Course", "id", courseId);
     }
 
     @Override
-    public String updateIsFinished(Integer courseId, boolean isFinished) {
-        Courses courses = coursesRepository.findById(courseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Courses", "id", courseId));
-
-        coursesRepository.switchFinished(courseId, isFinished);
-        return "SUCCESS";
+    public void updateIsFinished(Integer courseId, boolean isFinished) {
+        if (coursesRepository.existsById(courseId))
+            coursesRepository.switchFinished(courseId, isFinished);
+        else
+            throw new ResourceNotFoundException("Course", "id", courseId);
     }
 
+    /*TODO: calculate average review*/
     @Override
     public List<CourseReturnSearch> listAllCourseByKeyword(String keyword) {
         List<Courses> listCourses = coursesRepository.search(keyword);
 
-        return listCourses.stream().map(
-                courses -> {
-                    CourseReturnSearch response = modelMapper.map(courses, CourseReturnSearch.class);
-                    int totalReview = courses.getListReviews().size();
-                    int totalRating = courses.getListReviews().stream().mapToInt(Review::getRating).sum();
-                    double averageRating = (double) totalRating / totalReview;
-                    averageRating = Math.round(averageRating * 10.0) / 10.0;
-                    response.setAverageReview(averageRating);
-                    return response;
-                }
-        ).toList();
+        return listCourses.stream()
+                .map(courses -> modelMapper.map(courses, CourseReturnSearch.class))
+                .toList();
     }
 
-    private void convertSomeAttributeToEntity(Courses courses, CoursesRequest request) {
-        courses.setTitle(request.getTitle());
-        String slug = Utils.removeVietnameseAccents(request.getTitle());
-        courses.setSlug(slug);
-        courses.setDescription(request.getDescription());
-        courses.setPrice(request.getPrice());
-        courses.setDiscount(request.getDiscount());
-        courses.setEnabled(request.isEnabled());
-        courses.setPublished(request.isPublished());
-        courses.setFinished(request.isFinished());
-        if (request.isPublished()) {
-            courses.setPublishedAt(new Date());
-        }
-    }
-
-    private void sortChapterAndLesson(CourseResponse response) {
-        int totalLessonInCourse = 0;
-        List<ChapterDto> chapterList = response.getChapterList();
-        response.setTotalChapter(chapterList.size());
-        chapterList.sort(Comparator.comparingInt(ChapterDto::getOrders));
-        for (ChapterDto dto : chapterList) {
-            List<LessonResponse> listLesson = dto.getLessonList();
-            listLesson.sort(Comparator.comparingInt(LessonResponse::getOrders));
-            totalLessonInCourse += listLesson.size();
-            dto.setTotalLesson(listLesson.size());
-        }
-        response.setTotalLesson(totalLessonInCourse);
-    }
+//    private void sortChapterAndLesson(CourseResponse response) {
+//        int totalLessonInCourse = 0;
+//        List<ChapterDto> chapterList = response.getChapterList();
+//        response.setTotalChapter(chapterList.size());
+//        chapterList.sort(Comparator.comparingInt(ChapterDto::getOrders));
+//        for (ChapterDto dto : chapterList) {
+//            List<LessonResponse> listLesson = dto.getLessonList();
+//            listLesson.sort(Comparator.comparingInt(LessonResponse::getOrders));
+//            totalLessonInCourse += listLesson.size();
+//            dto.setTotalLesson(listLesson.size());
+//        }
+//        response.setTotalLesson(totalLessonInCourse);
+//    }
 
     private void sortChapterAndLesson(CourseReturnDetailPageResponse response) {
         int totalLessonInCourse = 0;
